Reorganized the files into a unified sublib
