+++ mala/descriptors/__init__.py
@@ -5,4 +5,3 @@ from .atomic_density import AtomicDensity
 from .descriptor import Descriptor
 from .minterpy_descriptors import MinterpyDescriptors
 from .ace import ACE
-from .ace_potential import ACEPotential
