+++ mala/descriptors/acelib/__init__.py
@@ -1,0 +1,3 @@
+"""Contains calculation/auxiliary functions for the ACE descriptors."""
+
+
