+++ mala/descriptors/acelib/ace_potential.py
@@ -3,7 +3,7 @@ import json
 
 import numpy as np
 
-import mala.descriptors.ace_coupling_utils as acu
+import mala.descriptors.acelib.ace_coupling_utils as ace_coupling_utils
 
 
 class ACEPotential:
@@ -97,7 +97,7 @@ class ACEPotential:
         else:
             nulst_1 = []
             for rank in self.ranks:
-                PA_lammps = acu.pa_labels_raw(
+                PA_lammps = ace_coupling_utils.pa_labels_raw(
                     rank,
                     nradmax_dict[rank],
                     lmax_dict[rank],
@@ -113,7 +113,7 @@ class ACEPotential:
         ns = []
         ls = []
         for nu in nus_unsort:
-            mu0ii, muii, nii, lii = acu.get_mu_n_l(nu)
+            mu0ii, muii, nii, lii = ace_coupling_utils.get_mu_n_l(nu)
             mu0s.append(mu0ii)
             mus.append(tuple(muii))
             ns.append(tuple(nii))
@@ -243,11 +243,13 @@ class ACEPotential:
             # betas = {ind:{nu:1.0 for nu in nulst} for ind in range(len(self.elements))}
             betas = {ind: {} for ind in range(len(self.elements))}
             for nu in nulst:
-                mu0, mu, n, l, L = acu.get_mu_n_l(nu, return_L=True)
+                mu0, mu, n, l, L = ace_coupling_utils.get_mu_n_l(
+                    nu, return_L=True
+                )
                 betas[mu0][nu] = 1.0
         for nu in nulst:
-            mu0, mu, n, l, L = acu.get_mu_n_l(nu, return_L=True)
-            rank = acu.get_mu_nu_rank(nu)
+            mu0, mu, n, l, L = ace_coupling_utils.get_mu_n_l(nu, return_L=True)
+            rank = ace_coupling_utils.get_mu_nu_rank(nu)
             try:
                 nus_per_rank[rank].append(nu)
             except KeyError:
